WIP: separate files, dir/single mode combine into prod_docs
