+++ lib/pbsimply/docengine/base.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby
+require 'pbsimply/docengine/pandoc.rb'
+require 'pbsimply/docengine/rdoc.rb'
+require 'pbsimply/docengine/misc.rb'
+require 'pbsimply/docengine/docutils.rb'
